requirements.txt
graphviz
python-jenkins
Flask==2.0.2
Flask-Cors==3.0.10
bash
curl
openssh==7.6
ip

